% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{database}
\alias{database}
\alias{db_create}
\title{Databases}
\arguments{
\item{name}{(character) A database name}

\item{...}{Further args passed on to \code{\link[httr]{POST}}}
}
\value{
Logical or a list, see Methods for what each returns
}
\description{
Databases
}
\section{Methods}{

\itemize{
 \item db_create: create a database
}
}
\examples{
\dontrun{
# make a client
(cli <- flux())

# make a database
cli$db_create(name = "thoughts")

# show a database
cli$db_show()
}
}

