% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flux}
\alias{flux}
\title{fluxy client}
\usage{
flux(host = "127.0.0.1", port = 8086, scheme = "http")
}
\arguments{
\item{host}{(character) Host url. Default: 127.0.0.1}

\item{port}{(character/numeric) Port. Default: 8086}

\item{scheme}{(character) http scheme, one of http or https. Default: http}

\item{path}{(character) URL path. Default: NULL/none}
}
\value{
Various output, see help files for each grouping of methods.
}
\description{
fluxy client
}
\details{
\code{flux} creates a R6 class object. The object is
not cloneable and is portable, so it can be inherited across packages
without complication.

\code{flux} is used to initialize a client that knows about your etcd
instance, with options for setting host, port, etcd api version,
whether to allow redirects, and the http scheme.
}
\section{flux methods}{

\strong{Methods}
  \describe{
    \item{\code{ping()}}{
     ping the InfluxDB server
    }
    \item{\code{version()}}{
     check the InfluxDB version
    }
    \item{\code{keys()}}{
     list keys, see also \code{\link{keys}}
    }
  }
}
\examples{
\dontrun{
# make a client
(cli <- flux())

# ping
## ping to make sure it's up
cli$ping()

# etcd variables
cli$host
cli$port
cli$scheme

# set a different host
flux(host = 'stuff.com')

# set a different port
flux(host = 3456)

# set a different http scheme
flux(scheme = 'https')
}
}

