% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxy-package.R
\docType{package}
\name{fluxy-package}
\alias{fluxy}
\alias{fluxy-package}
\title{fluxy, influxdb client}
\description{
fluxy, influxdb client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

