% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\title{Write data}
\arguments{
\item{name}{(character) A database name}

\item{data}{data to write}

\item{...}{Further args passed on to \code{\link[httr]{POST}}}
}
\value{
Logical or a list, see Methods for what each returns
}
\description{
Write data
}
\section{Methods}{

\itemize{
 \item write: create a database
}
}
\examples{
\dontrun{
# make a client
(cli <- flux())

# create a database
cli$db_create(name = "thoughts")

# write data
x <- 'cpu_load_short,host=server01,region=us-west value=0.64 1434055562000000000'
cli$write(name = "thoughts", data = x, config = verbose())
}
}

